First commit.
Here we introduce a parser just capable of recognising blocks of
sequential input: literals and variable insertions. The parser generates
s-expressions that are then fed into Template; they are kept separate so
we don't need to include the parser in a runtime library.

The Template does a simple tree walk evaluation and inserts variables by
looking them up in a 'scope' object.