Resolve promises bound to template variables.
To avoid needing to write code in your controller to wait for data to be
ready, the template language should resolve promises for you. When a
variable is looked up and resolved if necessary.

This is done simply: a template block is (currently) a sequence of
literals and insertions. We Promise.all() that sequence to wait for all
the chunks to be ready, then concat them.

This means Template.evaluate() must return a promise rather than a
string, since it may contain async code.