Accept readable streams as template input.
For a first attempt at stream support, we convert readable streams into
promises by piping into concat-stream to wait until they've emitted
everything.

This does produce correct output but is not ideal; we would rather have
templates stream out results as soon as they're available rather than
waiting for everything to be done, but that would require some
refactoring and a change to the Template.evaluate() interface.

For now, let's make this small change and then refactor in a later step.