Stream results out rather than waiting for completion.
In the last commit, we handled streams by wrapping them in a promise
that would wait for all the stream output before resolving. This blocks
the template from returning anything until all rendering is done.

Here, we turn this conversion inside-out: we wrap all output in a stream
that knows how to sequence everything.

Template chunks are now either a literal, or a variable that resolves to
one of:

- String
- Stream<String>
- Promise<String>

Evaluating a Template block produces an Output stream that wraps a list
of such values and queues them up for emission. Literals and strings are
emitted immediately. Streams are piped to the ouput and block further
emission until they end. Promises are unwrapped for emission and block
further emission until they're settled. Errors are emitted and halt
emission.

A mutex is needed around calls into _read() to prevent the queue being
modified while an item is being processed.