Fetch output of streams in parallel.
Previously, if you had a template like

    Hi {{ name }}, welcome to {{ city }}

then 'city' would not start being read until we reach that part of the
template, after 'name' has finished emitting.

By piping all the streams in the queue through a PassThrough, we get
them to start reading immediately into the PassThrough's buffer, ready
to be read when needed. This means each block renders as fast as its
single slowest stream, rather than as the sum of all of the streams.