Support foo.bar.baz style accessor chains.
This syntax evaluates each name in the chain, and if it's a promise,
unwraps it before looking up the next name. There's a couple of things
worth mentioning here.

The parser has to do some work to mangle the tree because PEG doesn't
support left-recursion. The JS property access operator (.) is
left-associative, that is 'foo.bar.baz' should parse as

    ((foo).bar).baz

But a PEG rule like

    access <- access "." varname

will recurse infinitely. So we use a trick; the parser rule is

    access_chain <- varname ("." varname)*

and the parser action _access_chain() restructures the resulting tree
into the left-associative structure

    ['access', ['access', ['name', 'foo'], 'bar'], 'baz']

The second thing is that nested promises are simply handled by the
_eval_access() function which evals a single instance of the '.'
operator. By wrapping a promise around the left-hand-side value we can
wait for that before looking up the next property. This promise becomes
the input for the next '.' in the chain, and the entire promise ends up
being handled by the Output stream and sequenced accordingly.