Introduce simple conditionals.
Here we introduce the syntax:

    {{ if condition }}
      block
    {{ end }}

'condition' can be any expression, which currently means variable
references or property access chains like 'foo.bar.baz'.

'block' can be any sequence of literal chunks and variable references,
but to prevent the parser treating '{{ end }}' as a variable, a
look-ahead is needed to disallow keywords as variable names. 'if' and
'end' are the only keywords.

The tree for this syntax looks like:

    ['cond', condition, block]

and the block is evaluated if the condition is true. Otherwise, an empty
string is emitted.

So far this only supports synchronous conditions, but streaming blocks
should work per existing functionality. It also does not support 'else'
or 'else-if' branches.